requirements.txt
# Core dependencies
PyQt5>=5.15.7
gTTS>=2.3.2
langdetect==1.0.9
fasttext-wheel>=0.9.2
mutagen>=1.45.1

# Web scraping dependencies
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.6.3

# Additional utilities
python-dotenv>=0.19.0
appdirs>=1.4.4 